==> Chapter 7/code/p0.1 install_packages.R <==
install.packages("devtools")
require(devtools)
devtools::install_version("mice", version = "2.46.0")
install.packages("foreach")
install.packages("doParallel")
install.packages("tidyverse")
install.packages("ggplot2")
install.packages("ggpubr")
install.packages("knitr")
